drf_cassandra_boilerplate/requirements.txt
cassandra-driver==3.6
Django==1.10.1
six
coverage
djangorestframework==3.4.6
django-filter==0.14.0
una-django-cassandra-engine==5.0.5
